Fix Grafana plugin validation: gosec G115 + govulncheck CVEs
- gosec G115 (26 HIGH): change Granularity.Size/DateTimeFormat.Size from
  uint to int64 (idiomatic for Go's int64 time math), removing the
  overflow-prone conversions. The one remaining Pinot INT->int32 narrowing
  is 32-bit by definition and annotated with a justified #nosec.
  Verified: `gosec -severity=high` -> 0 issues.
- govulncheck (27 stdlib CVEs): bump the go directive 1.25.7 -> 1.25.11,
  which is >= every fix version on the 1.25 branch (net/url, crypto/x509,
  net/textproto, Fchmodat, ...). CI's setup-go reads this directive, so
  release binaries build with 1.25.11. Verified: a 1.25.11 binary scans
  clean with `govulncheck -mode binary`.

Staying on Go 1.25.x (not 1.26) so Snyk can process go.mod; the SDK is left
at v0.292.0 since the only newer release (v0.292.1) requires Go 1.26.3+.
The README datetimeconvert link was already fixed on main.

+++ pkg/pinot/sql_expressions.go
@@ -95,19 +95,19 @@ func TimeFilterExpr(filter TimeFilter) SqlExpr {
 func TimeExpr(ts time.Time, format DateTimeFormat) SqlExpr {
 	switch format.Unit {
 	case TimeUnitNanoseconds:
-		return SqlExpr(fmt.Sprintf("%d", ts.UnixNano()/int64(format.Size)))
+		return SqlExpr(fmt.Sprintf("%d", ts.UnixNano()/format.Size))
 	case TimeUnitMicroseconds:
-		return SqlExpr(fmt.Sprintf("%d", ts.UnixMicro()/int64(format.Size)))
+		return SqlExpr(fmt.Sprintf("%d", ts.UnixMicro()/format.Size))
 	case TimeUnitMilliseconds:
-		return SqlExpr(fmt.Sprintf("%d", ts.UnixMilli()/int64(format.Size)))
+		return SqlExpr(fmt.Sprintf("%d", ts.UnixMilli()/format.Size))
 	case TimeUnitSeconds:
-		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/int64(format.Size)))
+		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/format.Size))
 	case TimeUnitMinutes:
-		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/int64(format.Size)/60))
+		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/format.Size/60))
 	case TimeUnitHours:
-		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/int64(format.Size)/3600))
+		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/format.Size/3600))
 	case TimeUnitDays:
-		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/int64(format.Size)/86400))
+		return SqlExpr(fmt.Sprintf("%d", ts.Unix()/format.Size/86400))
 	default:
 		return ""
 	}
